swap dedupe_v2 for dedupe
